requirements.txt
streamlit
requests
python-dotenv
sentence-transformers
faiss-cpu
numpy